+++ src/PeriodToMilliseconds.php
@@ -3,7 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\TimeWarp;
 
-use Innmind\TimeWarp\Exception\LogicException;
 use Innmind\TimeContinuum\Period;
 
 /**
@@ -12,13 +11,13 @@ use Innmind\TimeContinuum\Period;
 final class PeriodToMilliseconds
 {
     /**
-     * @throws LogicException If any number of months is specified
+     * @throws \LogicException If any number of months is specified
      */
     public function __invoke(Period $period): int
     {
         if ($period->months() !== 0) {
             // a month is not constant
-            throw new LogicException('Months can not be converted to milliseconds');
+            throw new \LogicException('Months can not be converted to milliseconds');
         }
 
         $second = 1000;