+++ CHANGELOG.md
@@ -10,3 +10,8 @@
 - `Innmind\TimeWarp\Halt` return an `Innmind\Immutable\Attempt<Innmind\Immutable\SideEffect>`
 - `Innmind\TimeWarp\PeriodToMilliseconds` is now internal
 - `Innmind\TimeWarp\Halt\Usleep` constructor is now private, use `::new()` named constructor
+
+### Removed
+
+- `Innmind\TimeWarp\Exception\Exception`
+- `Innmind\TimeWarp\Exception\LogicException`